+++ docs/changelog.md
@@ -37,6 +37,7 @@ See [upgrade notes][1] for helpful information when upgrading from previous vers
   - Azure Cache for Redis:
     - Check for legacy Azure Cache for Redis instances by @BenjaminEngeset.
       [#3605](https://github.com/Azure/PSRule.Rules.Azure/issues/3605)
+  - Azure Cache for Redis Enterprise and Enterprise Flash:
     - Check for deprecated Redis Enterprise and Enterprise Flash SKUs by @BenjaminEngeset.
       [#3606](https://github.com/Azure/PSRule.Rules.Azure/issues/3606)
   - Managed Instance for Apache Cassandra: